feat: make application scope and orchestrator url inputs optional
Add default values to the inputs orchestratorUrl an orchestratorApplicationScope optional.

The default value of orchestratorUrl assumes that the action will target an Orchestrator instance in Automation Cloud.

The default value of orchestratorApplicationScope adds the scopes specified in UiPath CLI documentation to reduce the risk of copy-paste errors from users utilzing the scopes as documented